fix: address self-review findings (keep nav server-rendered, isolate active-link state)
